+++ project/src/store/action.ts
@@ -1,7 +1,5 @@
 import { createAction } from '@reduxjs/toolkit';
 
-export const changeGenre = createAction('changeGenre', (updateGenre: string) => ({
+export const setGenre = createAction('changeGenre', (updateGenre: string) => ({
   payload: { updateGenre }
 }));
-
-export const getUpdatedFilmList = createAction('getUpdatedFilmList');
